large world map is added, will crash close to the edges if chunk is not multp. of 16

+++ Core/Source/Scene_Menu.cpp
@@ -112,12 +112,12 @@ void Scene_Menu::loadMenu(){
     entity1.getComponent<CTransform>().scale = Vec2{1.2f, 1.2f};
     entity1.addComponent<CName>("game_title");
 
-    spawnButton(Vec2 {128+0.0f,64*7+38.4f}, "button_unpressed", "new", "NEW GAME");
-    spawnButton(Vec2 {128+0.0f,64*9+12.8f}, "button_unpressed", "continue", "CONTINUE");
-    spawnButton(Vec2 {width()-196,64}, "button_unpressed", "720p", "720p");
-    spawnButton(Vec2 {width()-196,196}, "button_unpressed", "1080p", "1080p");
-    spawnButton(Vec2 {width()-196,320}, "button_unpressed", "1440p", "1440p");
-    spawnButton(Vec2 {width()-196,428}, "button_unpressed", "1440p Wide", "1440p Wide");
+    spawnButton(Vec2 {128.0f +0.0f,64*7.0f +38.4f}, "button_unpressed", "new", "NEW GAME");
+    spawnButton(Vec2 {128.0f +0.0f,64*9.0f +12.8f}, "button_unpressed", "continue", "CONTINUE");
+    spawnButton(Vec2 {(float)width()-196.0f,64.0f }, "button_unpressed", "720p", "720p");
+    spawnButton(Vec2 {(float)width()-196.0f,196.0f }, "button_unpressed", "1080p", "1080p");
+    spawnButton(Vec2 {(float)width()-196.0f,320.0f }, "button_unpressed", "1440p", "1440p");
+    spawnButton(Vec2 {(float)width()-196.0f,428.0f }, "button_unpressed", "1440p Wide", "1440p Wide");
 }
 
 void Scene_Menu::spawnLevel(const Vec2 pos, std::string level)
